chore: Move BUILDS_PAGE_SIZE
It was incorrectly under /types rather than /constants
