+++ src/routes/(main)/user/[name]/+page.svelte
@@ -7,7 +7,7 @@
   import type { FullStadiumBuild as Build } from "$lib/types/build";
   import { api } from "$lib/utils/api";
   import type { PageableBuildsSnapshot } from "$src/lib/types/snapshot";
-  import { BUILDS_PAGE_SIZE } from "$src/lib/types/page";
+  import { BUILDS_PAGE_SIZE } from "$lib/constants/page";
 
   const { data } = $props();
