+++ src/routes/api/builds/latest/+server.ts
@@ -1,6 +1,6 @@
 import { prisma } from "$src/database/prismaClient.server.js";
 import { FullStadiumBuildInclude } from "$src/lib/types/build";
-import { BUILDS_PAGE_SIZE } from "$src/lib/types/page.js";
+import { BUILDS_PAGE_SIZE } from "$lib/constants/page";
 import { headers } from "$src/lib/constants/api";
 
 export async function GET({ url }) {
