+++ src/routes/api/builds/user/[id]/+server.ts
@@ -1,5 +1,5 @@
 import { prisma } from "$src/database/prismaClient.server.js";
-import { BUILDS_PAGE_SIZE } from "$src/lib/types/page.js";
+import { BUILDS_PAGE_SIZE } from "$lib/constants/page";
 import { headers } from "$src/lib/constants/api";
 
 export async function GET({ url, params }) {
